added realm  selection

+++ scripts/stop.sh
@@ -25,7 +25,10 @@ if [[ ! " ${configs[@]} " =~ " ${ENVIRONMENT} " ]]; then
   exit 1
 fi
 
-echo "Deleting volumes and containers for environment: $ENVIRONMENT"
-#echo "Remove -v flag from scripts/stop.sh to keep volumes."
-docker-compose --project-directory . --env-file "./infrastructure/docker-compose/${ENVIRONMENT}/.env" -f "./infrastructure/docker-compose/${ENVIRONMENT}/docker-compose.yaml" down -v
+if [[ $2 == "clean" ]]; then
+  echo "Deleting volumes and containers for environment: $ENVIRONMENT"
+  docker-compose --project-directory . --env-file "./infrastructure/docker-compose/${ENVIRONMENT}/.env" -f "./infrastructure/docker-compose/${ENVIRONMENT}/docker-compose.yaml" down -v
+  exit 0
+fi
+docker-compose --project-directory . --env-file "./infrastructure/docker-compose/${ENVIRONMENT}/.env" -f "./infrastructure/docker-compose/${ENVIRONMENT}/docker-compose.yaml" down
 exit 0
